basic compile works
