+++ visivo/query/jobs/run_trace_job.py
@@ -2,14 +2,12 @@ from visivo.logging.logger import Logger
 from visivo.models.base.parent_model import ParentModel
 from visivo.models.model import CsvScriptModel, Model
 from visivo.models.target import Target
-from visivo.query.runner import Runner, format_message
+from visivo.query.jobs.job import format_message
 from time import time
 
 
 def action(trace, dag, output_dir, errors):
-    model = ParentModel.all_descendants_of_type(
-        type=Model, dag=dag, from_node=trace
-    )[0]
+    model = ParentModel.all_descendants_of_type(type=Model, dag=dag, from_node=trace)[0]
     if isinstance(model, CsvScriptModel):
         target = model.get_target(output_dir=output_dir)
     else:
@@ -35,7 +33,9 @@ def action(trace, dag, output_dir, errors):
                 status=f"\033[32mSUCCESS\033[0m {round(time()-start_time,2)}s",
                 full_path=trace_query_file,
             )
-            Runner.aggregate_data_frame(data_frame=data_frame, trace_dir=trace_directory)
+            Runner.aggregate_data_frame(
+                data_frame=data_frame, trace_dir=trace_directory
+            )
             Logger.instance().success(success_message)
         except Exception as e:
             failure_message = format_message(
